+++ packages/fiddle/src/assets.d.ts
@@ -3,3 +3,5 @@ declare module '*.png' {
   const src: string;
   export default src;
 }
+
+declare module '*.css';